refactor: remove not used utils
