+++ src/consumer/config/component-config.ts
@@ -4,7 +4,7 @@ import { BitId } from '../../bit-id';
 import { DEFAULT_REGISTRY_DOMAIN_PREFIX } from '../../constants';
 import ShowDoctorError from '../../error/show-doctor-error';
 import logger from '../../logger/logger';
-import filterObject from '../../utils/filter-object';
+import { filterObject } from '../../utils';
 import Component from '../component/consumer-component';
 import PackageJsonFile from '../component/package-json-file';
 import AbstractConfig from './abstract-config';
